build: refactor build pipeline
This PR removes the need for the custom build scripts living in ./bin

AFAIK this is a feature complete replacement except for one change.
The input point of the app has been modified to lib/binaural-fir.js and
the babelify transform has been added (and included in the package.json).

If someone is requiring this package in a browserify workflow the transform will
automagically be applied to the subtree when required.

I'm not 100% you will like all of these changes but figured it is a good place to start.

The script ```npm run build``` will now build all the assets.
