--- binaural-fir.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/binaural-fir');
